Add a new overview topic to fix #103
Creates to-do-overview_allenaavila.md.
Adds a new and improved to-do overview topic.
Updated index with link.
Updated overview topic to comply with Vale and markdownlint specifications.
Update to-do-overview_allenaavila.md

Updated the content of to-do-overview_allenaavila.md to meet the requirements from the markdownlint and Vale tools.

Update To-Do Service to match current style